improve protein image classifier
