--- Stanford360/Protein/Model/ObjectDetector.mlmodel.license
@@ -1,7 +1,0 @@
-
-This source file is part of the Stanford 360 based on the Stanford Spezi Template Application project
-
-SPDX-FileCopyrightText: 2025 Stanford University
-
-SPDX-License-Identifier: MIT
-
