+++ Stanford360/Protein/View/AddMealView.swift
@@ -50,12 +50,13 @@ struct AddMealView: View {
                 sourceSelectionButtons
             }
             .onChange(of: selectedImage) { _, newImage in
-                // classifier.image = newImage
-                if let image = newImage {
-                    classification(image: image)
-                } else {
-                    print("No image selected")
-                }
+                classifier.image = newImage
+                // classifier.classifyImage(newImage)
+//                if let image = newImage {
+//                    classification(image: image)
+//                } else {
+//                    print("No image selected")
+//                }
             }
             .onChange(of: highestConfidenceClassification) { _, newValue in
                 if let classification = newValue, !classification.isEmpty {
@@ -244,81 +245,81 @@ extension AddMealView {
 }
 
 
-extension AddMealView {
-    // convert A image to cv Pixel Buffer with 224*224
-    func convertImage(image: UIImage) -> CVPixelBuffer? {
-        let newSize = CGSize(width: 224.0, height: 224.0)
-        UIGraphicsBeginImageContext(newSize)
-        image.draw(in: CGRect(origin: CGPoint.zero, size: newSize))
-        
-        guard let resizedImage = UIGraphicsGetImageFromCurrentImageContext() else {
-            return nil
-        }
-        
-        UIGraphicsEndImageContext()
-        
-        let attributes = [
-            kCVPixelBufferCGImageCompatibilityKey: kCFBooleanTrue,
-            kCVPixelBufferCGBitmapContextCompatibilityKey: kCFBooleanTrue
-        ] as CFDictionary
-        var pixelBuffer: CVPixelBuffer?
-        let status = CVPixelBufferCreate(
-            kCFAllocatorDefault,
-            Int(newSize.width),
-            Int(newSize.height),
-            kCVPixelFormatType_32ARGB,
-            attributes,
-            &pixelBuffer
-        )
-        
-        guard let createdPixelBuffer = pixelBuffer, status == kCVReturnSuccess else {
-            return nil
-        }
-        
-        CVPixelBufferLockBaseAddress(createdPixelBuffer, CVPixelBufferLockFlags(rawValue: 0))
-        let pixelData = CVPixelBufferGetBaseAddress(createdPixelBuffer)
-        
-        let colorSpace = CGColorSpaceCreateDeviceRGB()
-        guard let context = CGContext(
-            data: pixelData,
-            width: Int(newSize.width),
-            height: Int(newSize.height),
-            bitsPerComponent: 8,
-            bytesPerRow: CVPixelBufferGetBytesPerRow(createdPixelBuffer),
-            space: colorSpace,
-            bitmapInfo: CGImageAlphaInfo.noneSkipFirst.rawValue
-        ) else {
-            return nil
-        }
-        
-        context.translateBy(x: 0, y: newSize.height)
-        context.scaleBy(x: 1.0, y: -1.0)
-        
-        UIGraphicsPushContext(context)
-        resizedImage.draw(in: CGRect(x: 0, y: 0, width: newSize.width, height: newSize.height))
-        UIGraphicsPopContext()
-        CVPixelBufferUnlockBaseAddress(createdPixelBuffer, CVPixelBufferLockFlags(rawValue: 0))
-        
-        return createdPixelBuffer
-    }
-}
+// extension AddMealView {
+//    // convert A image to cv Pixel Buffer with 224*224
+//    func convertImage(image: UIImage) -> CVPixelBuffer? {
+//        let newSize = CGSize(width: 224.0, height: 224.0)
+//        UIGraphicsBeginImageContext(newSize)
+//        image.draw(in: CGRect(origin: CGPoint.zero, size: newSize))
+//        
+//        guard let resizedImage = UIGraphicsGetImageFromCurrentImageContext() else {
+//            return nil
+//        }
+//        
+//        UIGraphicsEndImageContext()
+//        
+//        let attributes = [
+//            kCVPixelBufferCGImageCompatibilityKey: kCFBooleanTrue,
+//            kCVPixelBufferCGBitmapContextCompatibilityKey: kCFBooleanTrue
+//        ] as CFDictionary
+//        var pixelBuffer: CVPixelBuffer?
+//        let status = CVPixelBufferCreate(
+//            kCFAllocatorDefault,
+//            Int(newSize.width),
+//            Int(newSize.height),
+//            kCVPixelFormatType_32ARGB,
+//            attributes,
+//            &pixelBuffer
+//        )
+//        
+//        guard let createdPixelBuffer = pixelBuffer, status == kCVReturnSuccess else {
+//            return nil
+//        }
+//        
+//        CVPixelBufferLockBaseAddress(createdPixelBuffer, CVPixelBufferLockFlags(rawValue: 0))
+//        let pixelData = CVPixelBufferGetBaseAddress(createdPixelBuffer)
+//        
+//        let colorSpace = CGColorSpaceCreateDeviceRGB()
+//        guard let context = CGContext(
+//            data: pixelData,
+//            width: Int(newSize.width),
+//            height: Int(newSize.height),
+//            bitsPerComponent: 8,
+//            bytesPerRow: CVPixelBufferGetBytesPerRow(createdPixelBuffer),
+//            space: colorSpace,
+//            bitmapInfo: CGImageAlphaInfo.noneSkipFirst.rawValue
+//        ) else {
+//            return nil
+//        }
+//        
+//        context.translateBy(x: 0, y: newSize.height)
+//        context.scaleBy(x: 1.0, y: -1.0)
+//        
+//        UIGraphicsPushContext(context)
+//        resizedImage.draw(in: CGRect(x: 0, y: 0, width: newSize.width, height: newSize.height))
+//        UIGraphicsPopContext()
+//        CVPixelBufferUnlockBaseAddress(createdPixelBuffer, CVPixelBufferLockFlags(rawValue: 0))
+//        
+//        return createdPixelBuffer
+//    }
+// }
 
 extension AddMealView {
-    func classification(image: UIImage) {
-        guard let mobilenet = try? MobileNetV2(configuration: MLModelConfiguration()) else {
-            print("ERROR: Failed to initialize MobileNetV2 model")
-            return
-        }
-        if let imagebuffer = convertImage(image: image) {
-            do {
-                let prediction = try mobilenet.prediction(image: imagebuffer)
-                print(prediction.classLabel)
-                highestConfidenceClassification = prediction.classLabel
-            } catch {
-                print("ERROR prediction ", error)
-            }
-        }
-    }
+//    func classification(image: UIImage) {
+//        guard let mobilenet = try? MobileNetV2(configuration: MLModelConfiguration()) else {
+//            print("ERROR: Failed to initialize MobileNetV2 model")
+//            return
+//        }
+//        if let imagebuffer = convertImage(image: image) {
+//            do {
+//                let prediction = try mobilenet.prediction(image: imagebuffer)
+//                print(prediction.classLabel)
+//                highestConfidenceClassification = prediction.classLabel
+//            } catch {
+//                print("ERROR prediction ", error)
+//            }
+//        }
+//    }
     
 //    func processClassificationResults(_ results: [VNClassificationObservation]) {
 //        let validResults = results.filter { $0.confidence >= 0.5 }
