+++ Stanford360/Protein/View/ProteinView.swift
@@ -30,6 +30,7 @@ struct ProteinView: View {
 					ScrollView {
 						mealsCardView
 					}
+                    .frame(maxWidth: .infinity, alignment: .leading)
 					.padding(20)
 				}
